"[TeamCity] OEMx PantherLake_pantherlake_family_Daily Daily Updating Version/ID: PantherLake_2342_00
Hsd-es-id: N/A"
Original commit date: Tue Aug 20 04:32:26 2024 +0000
Original commit hash: 6553dead7eeb10131281292ddec3aa7cf45bc2e0

+++ OneSiliconPkg/IncludePrivate/Register/Ptl/Sncu/SncuCfg.h
@@ -1524,7 +1524,7 @@ typedef union {
                                of bits [3:1] in this register.
                             */
 
-    UINT32 admsk1024 : 1;
+    UINT64 admsk1024 : 1;
 
                             /* Bits[30:30], Access Type=RW, default=0x00000000*/
